Clean up PDP11 sign comparisons
Use an int32 temporary for the RP11 cylinder update path instead of
reusing the unsigned sector-count variable. This keeps the temporary in
the same domain as the surrounding cylinder and RPCA values.

Also make the RH11 write-check comparison explicit by comparing the
PDP-11 memory word as an unsigned value against the Massbus data buffer.

This fixes the targeted -Wsign-compare warnings without changing the
guest-visible behavior.

+++ simulators/PDP11/pdp11_rh.c
@@ -641,7 +641,7 @@ for (i = 0; i < bc; i = i + pbc) {                      /* loop by pages */
         pbc = bc - i;
     for (j = 0; j < pbc; j = j + 2) {                   /* loop by words */
         massbus[mb].db = *buf++;                        /* get dev word */
-        if (RdMemW (pa) != massbus[mb].db) {            /* miscompare? */
+        if ((uint32) RdMemW (pa) != massbus[mb].db) {   /* miscompare? */
             mba_set_cs2 (CS2_WCE, mb);                  /* set error */
             massbus[mb].cs3 = massbus[mb].cs3 |         /* set even/odd */
                 ((pa & 1)? CS3_WCO: CS3_WCE);
